Add AuthenticationController
The AuthenticationController is used to generate a LongLivedToken token which is hashed and stored in the DB, same as the TemporaryToken.
It does so by receiving a TemporaryToken, checking for validity, expiration, etc... if said token is valid a LongLivedToken is created and returned to the player.
This LongLived Token will is needed to play the game.