Add GuessesController
The GuessesController is used to create guesses based on coordinates supplied by the player.
We do this using the `interactor` gem to create a service layer to encapsulate the business logic.
It works in the order as follows as organized in the PlayGame organizer:
AuthenticatePlayer - checks if the longlivedtoken supplied is valid, returns an unauthorized error if not.
CalculateDistance - using googles api via the `geokit` gem we calculate the distance of the coordinates provided, we return an error not_found if no treasures currently exist.
CreatePlayerGuess - creates a Guess record for the player, it returns an error if the guess cannot be created, and sets messages for the guesses outcome (winner or loser)
EmaileWinner - if the guess is correct, we then send a confirmation email to the player.